+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.0] - June 24, 2021
+
+* Docking layout
+  * Docking areas
+* Initial version of drag and drop areas
+
 ## [0.0.1] - June 20, 2021
 
 * Package creation